fix: simplify compaction advancement classifier
