+++ test/compaction-advance.test.mjs
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { didConversationAdvance } from "../dist/compaction-advance.js";
+import { didConversationAdvance } from "../dist/index.js";
 
 const anchor = { id: "compact-turn", type: "message", message: { role: "assistant" } };
 
@@ -9,11 +9,13 @@ function branchWith(...tail) {
     return [{ id: "root", type: "message", message: { role: "user" } }, anchor, ...tail];
 }
 
-test("allows only passive extension metadata", () => {
+test("allows non-contextual session entries", () => {
     const branch = branchWith(
         { id: "extension-state", type: "custom" },
         { id: "checkpoint", type: "label" },
         { id: "session-name", type: "session_info" },
+        { id: "model", type: "model_change" },
+        { id: "thinking", type: "thinking_level_change" },
     );
 
     assert.equal(didConversationAdvance(branch, anchor.id), false);
@@ -39,9 +41,17 @@ test("cancels when any message or contextual custom message arrives", () => {
     );
 });
 
-test("cancels for other state-changing entries or a missing compact anchor", () => {
+test("cancels for contextual entries, unknown types, or a missing compact anchor", () => {
     assert.equal(
-        didConversationAdvance(branchWith({ id: "model", type: "model_change" }), anchor.id),
+        didConversationAdvance(branchWith({ id: "branch-summary", type: "branch_summary" }), anchor.id),
+        true,
+    );
+    assert.equal(
+        didConversationAdvance(branchWith({ id: "compaction", type: "compaction" }), anchor.id),
+        true,
+    );
+    assert.equal(
+        didConversationAdvance(branchWith({ id: "future", type: "future_context" }), anchor.id),
         true,
     );
     assert.equal(didConversationAdvance(branchWith(), "missing"), true);
